feat(build): upgrade Maven shade plugin version and refactor main method
I have upgraded the version of the Maven shade plugin in the pom.xml file to 3.5.1. Additionally, I have refactored the main method in the CapitalHubApp.java file to call the main method in the CapitalHubApp class directly.

This commit improves the build process and simplifies the entry point of the application.