- refactor(action): make name and symbol fields final, refactor(ActionProduit): remove final modifier from quantity field, refactor(APIFinnhubFetcher): initialize apiURL with an empty string, style(BuySellForm): remove fx:controller attribute, move(controller): move BuySellController to BuySellFormController, style(CapitalHubApp): reduce height of primaryStage, chore(ConnectionForm): remove unused imports, refactor(DataBaseConnectionSingleton): remove InstanceHolder class, chore(DatabaseFeeder): remove unused code, fix(History): fix save method to handle insert and update, chore(InscriptionForm): fix formatting in fxml, fix(Insertion.sql): correct values for Buy and Sell actions - fix(db): fix typo in transaction type values
The commit fixes a typo in the transaction type values in the database. The word 'Buy' was changed to 'BUY' and 'Sell' was changed to 'SELL'.

The commit also includes changes in the ListActionController.java file, where imports related to the User class were removed, and an import for the javafx.scene.Parent was added. Additionally, code related to the main method was removed, a refresh method was added, and changes were made in the actionBuySell method to handle exceptions and load the BuySellForm.fxml file.

The ListActions.fxml and ListPortefeuille.fxml files were modified to remove unnecessary code.

In the ListPortefeuilleController.java file, changes were made to the makeTransaction method to update the quantity of the actionProduit and add a new History object.

In the PieChartController.java file, a missing closing bracket was added.

In the Portefeuille.java file, changes were made to the id, name, and description variables.

In the Transaction.java file, changes were made to the id, save, and insert methods.

In the TransactionType.java file, a toString method was added to override the default toString behavior.

In the User.java file, changes were made to the id and getUsername methods.